ADD Schemas do FLASK_SQLALCHEMY

+++ src/models/character_episode_model.py
@@ -3,8 +3,8 @@ from configs.database import db
 class CharacterEpisode(db.Model):
     __tablename__ = 'character_episodes'
     
-    character_id = db.Column(db.BigInteger, db.ForeignKey('character.id'), primary_key=True)
-    episode_id = db.Column(db.BigInteger, db.ForeignKey('episodes.id'), primary_key=True)
+    character_id = db.Column(db.Integer, db.ForeignKey('character.id'), primary_key=True)
+    episode_id = db.Column(db.Integer, db.ForeignKey('episodes.id'), primary_key=True)
 
 ###
 #CREATE TABLE character_episodes (
